Fix _parse_response in the case that self._key is not a list
In the case where only one object of a key is returned, tradier does not wrap the object in a list (as in the case when many objects are returned). This is a byproduct of the json format mentioned here: https://documentation.tradier.com/brokerage-api/overview/response-format.

For example, issue is most obvious in the lookup file. If a ticker is looked up, and only one ticker is returned (ex GOOGL), self._key is just a json object. However, _parse_response expects self._key to be a list of objects, resulting in an AttributeError: 'str' object has no attribute 'keys'.

The solution is to simply wrap self._key in a list if self._key happens not to be a list.